Remove duplication from DSI API to spreadsheet tests
Introduce shared examples to remove duplication from add_dsi_approvers_to_spreadsheet and add_dsi_users_to_spreadsheet usecases
